Convert ScopeManager to interface, add ownership semantics
The previous idea was to have ScopeManager as a Utility like class with the
thread_local implementation. After reviewing the Java implementation, I have
made the ScopeManager an interface and a seperate ThreadLocalScopeManager
class will provide the thread_local implementation.

The reference based interface has been replaced with shared_ptr. This is to
allow the Span to be consumed from the ScopeManager and have its lifetime
managed by the application/library code as appropriate. E.g. consider an
application that keeps the Span alive until an Async signal is received.