fix: use std::invoke_result instead of std::result_of
std::result_of was deprecated in c++17 and removed in c++20. This commit
configures the project to use c++17 if the compiler supports it and use
std::invoke_result instead of std::result_of.

If the compiler doesn't support c++17, the project is configured to fallback to
c++11 and still use std::result_of.